[check-licenses] Support target command
Reland fxr/528786.

The previous attempt had a few bugs that caused problems:
1. Most egregiously the output code had a break in the wrong place,
which caused the used/not used determination to be completely wrong.

2. There was an existing problem where when a Match was merged, the old
object was returned - so any modifications done to it were discarded.

3. Single license files are not marked as used unless a file is
processed that uses that project license. Some directories that contain
single licenses are skipped, even though the license should be included.
As a workaround for now we can mark every single license file as used.
In the future, we should either start processing the skipped directories
so that the GN dependency analysis can be done, or do a second pass for
skipped directories to determine whether or not there is a dependency.

Change-Id: I10bb60d2471cf503b7aaeac6900db62431747109
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541164
Reviewed-by: Jerry Belton <jcecil@google.com>
Commit-Queue: Alex Legg <alexlegg@google.com>