[fidl][llcpp] WireSharedClient: Rename Unbind to AsyncTeardown.
Unbind is understood as extracting the channel from the client and
providing it back to the user. This doesn't match the current semantics
in LLCPP, which starts the asynchronous procedure of shutting down the
client. "Teardown" is a more accurate name.

I considered naming it |AsyncClose|, but it's not as accurate as the
client owns the channel. For example, the channel will not be closed
if there are hanging synchronous calls, independent of the unbinding
process.

In the future, we may be able to re-introduce an asynchronus
`void AsyncUnbind(fit::callback<void(fidl::ClientEnd<Protocol>)>)`
that better matches user expectations.

Also updates the documentation of WireSharedClient to talk about the
two-phase shutdown pattern.

Bug: 75485
Change-Id: I03dd07d680650dae314b5a46ed288c856af5a655
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539462
Commit-Queue: Yifei Teng <yifeit@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
API-Review: Ian McKellar <ianloic@google.com>