[fx][daemon] Add `ffx daemon log --follow`
This allows the ffx daemon log to be followed for changes. In addition,
this also tweaks how to print the last few lines of the log file in
order to not buffer the entire log file in memory, and makes sure that
both features work together.

Bug: 75500
Change-Id: I389bdb41547a7216083230b7c6644a75efcea6ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541982
Reviewed-by: Jordon Wing <jwing@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>