[bt][rfcomm] Implement RfcommTest FIDL API
Update RFCOMM Session logic to support receiving the test requests,
and update handlers to expect appropriate responses from the remote
peer.

For the Disconnect request, we expect either a UA or DM response.
In either case, the RFCOMM Session with the peer will be terminated.

For the RLS request, we can only send it for established channels. In
this case, the peer is expected to respond by echoing the command with
the same parameters.

Bug: 61306
Test: fx test bt-rfcomm-tests
Test: PTS RFCOMM/DEVA-DEVB/RFC/BV-04-C, RFCOMM/DEVA-DEVB/RFC/BV-14-C

Change-Id: I5b500ff8e534deafe58f85aa8d3e8a7bc989fb06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540942
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>