[fvm] Rename version variables.
Brings the naming of the FVM variables and constants in line with the
current policy. The old policy had a "format version" and a "software
revision".

The format version is the same as the major version. This was renamed
which should have no effect.

The software revision has been changed to minor version. In the old
scheme the software revision was monotonically increasing, while minor
versions reset for major version changes. However, since we have never
changed the major version, this change will also have no practical
effect.

No changes to on-disk formats or any values stored in them have been
made. Only the internal names in the software and some metrics cales
have change to be more consistent, and the policy for future
modifications has changed.

Change-Id: I914364b5b4d6483814af170d6fef2f9c8d5fb8b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543244
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>