[netemul] Use relative URLs
Wherever possible in netemul, replace absolute static URLs with relative
ones so that the netemul sandbox can be embedded in test packages. It's
currently not possible to use relative URLs with RealmBuilder, so that
will come in a future change.

Also, remove the netemul sandbox package and leave it as a plain
component so that it can be embedded in test packages.

Bug: 78313
Test: fx test -d netstack-integration-tests-v2
Change-Id: I281418d1c73a123b936eb6a1dadd67fb18aa57df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540722
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>