[SetUI][FIDL] Use jobs for Factory Reset Interface.
This changelist replaces the existing fidl processor-based
implementation of the FIDL Factory Reset Interface with
a version based on the new job infrastructure.

This is a reland of https://fxrev.dev/523986 and includes an additional
fix to avoid a crash on bad input.

Bug: 70534
Fixed: 78846
Test: fx test -o setui_service_tests
Change-Id: Ifebfad2720aa1f492eb567c20e49b2bd8d0d02f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544064
Reviewed-by: Sijie Chen <sijiec@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Commit-Queue: Paul Faria <paulfaria@google.com>