[fx] Ensure ssh is taken down for --no-serve
The prior code waited for a Ctrl-C, but because the signal handler
was local, only the local ssh connections were torn down, leaving
the remote ones running. Instead, here we use a remote wait with
a pseudo ttl, so the signal is trapped by the remote host, taking
down the remote connections.

Bug: 78388
Change-Id: I44a7f32d984699fc4c983ca287e7b1a2051ab665
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542395
Commit-Queue: Gary Boone <gboone@google.com>
Reviewed-by: Mahesh Saripalli <maheshsr@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>