[flutter] do not merge cml files in build step
In preparation for starting work on the CFv2 version of the
flutter and dart runners we need to have flutter and dart
components to launch. The current build rules attempt to merge
the cmx file with a file provided by the runner. This breaks
when we are trying to build a v2 component so we will just skip
merging the files for cml files. In the future we will need to
either include a shard.cml file or do some other type of merge
in this step but for now we just leave the manifest as is.

Change-Id: If307aac2d517bac109c41565f35d1413547f7277
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545041
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>