[kernel][hypervisor][arm64] Avoid reuse of hypercall numbers.
EL2 currently has two modes: a hypervisor enabled mode (where hypercalls
such as "guest resume" are available to the host), and a disabled mode
(where such hypercalls are not available).

In the current setup, hypercall 3 is allocated to the "on" hypercall
when the hypervisor is disabled, and "off" when it is enabled.

This can lead to problems if host code (due to its own bug) gets
confused as to the state of the hypervisor. In particular, the "on"
hypercall expects a stack pointer for EL2 to use, while the "off"
hypercall has no parameters. A kernel thread inadvertently calling "on"
when it expects to call "off" will start using whatever value happens to
be in the register `x1` as its stack.

We update the hypercall numbers to avoid reuse, and update the invalid
hypercalls to return ZX_ERR_BAD_STATE.

Bug: 77932
Change-Id: I6a2b59296230a7d2e9aa88e3bbc43d9c430273db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544302
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>