[acpica] Prepare to move acfuchsia.h in-tree.
Currently the two parts of Fuchsia's OSL are in wildly separate parts of
the tree, and changing them together requires a hard transition.

This CL adds a copy of acfuchsia.h inside the tree, so that a future CL
on //third_party/acpica can remove it from there.

Bug: 79000
Change-Id: Ida71cb27e5f5c6989c3e14d12cbeec29b6b56d1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545142
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>