[audio_core] Stream ramping and gain-limiting
This CL is needed to minimize/eliminate audible 'pops', and to
correct the frequency spectrum for Communications audio.

When calculating a MixJob's applied_gain, include each stream's
stream-specific gain (including gain-ramping).

The MixStage maintains an 'applied_gain_db' value for each MixJob,
representing "the total amount of gain applied to the resulting
output stream". This is sent to downstream EffectStages as an
inexpensive estimate of the mix stream's magnitude, and is used
by the MixStage itself to determine whether it has produces only
silence and thus can inform downstream MixStages of this.

To determine 'applied_gain_db' for a MIX of streams, we use the
largest gain of any of the sources to that mix. When doing this,
previously we only included the "shared" (i.e. usage) gain factor
in this calculation, trying to prevent any single stream from
unduly influencing the resultant value. However, a single stream
CAN distort the resultant mix, and including stream-specific
gain factors is more true to the intent of this mechanism. If a
stream is gain-ramping, we use the maximum gain-scale value found
in the array of gain values.

One important outcome of this change is that we now correctly
incorporate min_gain_db/max_gain_db gain-limits that have been
specified in audio_core_config.json for that MixStage. This fixes
incorrect behavior in effects like SaturatedGain, which limits
itself such that the gain it contributes, plus "system_volume"
(the applied_gain_db we pass in) never exceeds 0dB. This in turn
is needed in order to meet the parity requirements for the audio
spectrum for communications streams.

To avoid breaking tests in other repos, this change has been
staged in multiple CLs in a certain sequence. As a side effect,
the kFIdelityDbTolerance constant is temporarily increased to 2db.

Test: audio_core_unittests audio-core-pipeline-tests

BUG: b/186799991
BUG: 78862

Change-Id: I9c0e52385a0f8e770bfbe7d7cfd3e42eb1340a0d