[virtcon2] Add dpi argument to virtcon2
This boot argument can be used to specify DPI buckets that
are used to determine the scale factor that impacts
font size. The value should be set to a comma seperated
set of values that determine what the DPI buckets are
for scale factors 1x, 2x, 3x, etc. The lowest value
determine minimum DPI for 2x scale factor, the second lowest
determine minimum DPI for 3x scale factor, and so on.

For example, if we want 2x scale factor above 160 DPI and
a 3x scale factor above 320 DPI, then we can set virtcon.dpi
to "160,320".

The default is an empty list, which means that the scale
factor will always be 1.0;

Change-Id: I0967b8eacd744e15f14308e1a2942acbd5e52df1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545462
Reviewed-by: Rob Tsuk <robtsuk@google.com>
Commit-Queue: David Reveman <reveman@google.com>