[rbe] Avoid listing unstable directories as inputs
reclient assumes that the contents of `--input` directories are stable.
Allow-listing files is always preferred and safer.
Attempt to snapshot the files that are intended: .a, .o, .so (these are
not temporary, and will not disappear randomly).

There can still be instability if multiple concurrent build actions
write outputs to the same directory.
Adding files to the same dir could reduce a cache-hit rate,
and removing (temporary) files could cause the reproxy process
to fail with "file not found."

It is safe to over-specify inputs.
If more precision is desired in the future, then -Lnative flags will need
to be examined.

Change-Id: I56c650c4c18e04401f1fec9f54292e3bd1bb1b94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544221
Commit-Queue: David Fang <fangism@google.com>
Fuchsia-Auto-Submit: David Fang <fangism@google.com>
Reviewed-by: Shai Barack <shayba@google.com>