[build] Specify path to asan libclang_rt relative to the root build dir
When building for asan, we pass a path to the appropriate clang runtime
support library as a rustflags argument. The path is computed by running
the clang binary which outputs an absolute path. This rebases the path
to be relative to the root build dir per our best practices:
https://fuchsia.dev/fuchsia-src/concepts/build_system/best_practices?hl=en#prefer-relative-paths-from-rebase-path

Change-Id: I6eb49d977d36eac22d5b63f6d94664accfdac526
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544181
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Shai Barack <shayba@google.com>