[fxfs] Change the flush strategy and block-size
This changes the flush strategy for object stores so that we will
eventually do full compactions.

To get it it to work, I had to increase the block size for persistent
layers and object stores as otherwise I was finding that flushing ended
up using up so much journal space that it would trigger an immediate
flush again which led to an infinite loop.

I also add to address a TODO in the allocator so that now we'll sync the
journal and flush the device to make available pending deallocations.

Change-Id: Ic5a03f502e3eac23d2444ea2385a95b60245714d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541884
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>