[fxfs] Clean up on major compaction
During a major compaction (one which goes down to the base layer), we
have an opportunity to remove records that no longer need to be kept
around, such as tombstone records and extent deletions.

We need to do this as a second pass after the compacted layers are
merged, rather than during merging, since otherwise we might discard a
record that should have applied over another (e.g. an overlapping extent
deletion and allocation). This is implemented with a
MajorCompactionIterator that wraps the regular merging iterator and
discards records when possible.

This also includes a bug fix: When tombstoning files, we didn't actually
delete all of the extents. Fixed and added a test.

Tested: Unit test.
Change-Id: Iad9dd27e19f65c38426ccf9d99f533141866d8f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542881
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>