[fidlc] Parse modifiers into new raw::Modifiers AST node in new syntax
There are two reasons for this change:
    * It enables new lints that check the order of modifiers.
    * It moves some decisions about the meaning of modifiers (for
    example, what is the default strictness of some layout?) to the
    compiler, where they seem to be a conceptually better fit than the
    parser.

Of note, this CL has to account for an edge case like:

    struct { foo flexible; };

This is done by modifying the ParseCompoundIdentifier() method to take
the first Identifier as an argument, so that we may convert a
maybe_modifier of "flexible" in this position into the identifier for
a NamedLayoutReference.  Because we have to "lookahead" to deduce what
the last token in a set of possible modifiers means, and could
possibly produce multiple raw AST nodes when doing modifier parsing, the
old ParseModifiers() method is removed, and modifier parsing is done
inside of the ParseTypeConstructorNew() method.

Test: fidldev test fidlc
Bug: 72121
Change-Id: I561c8064b57921263009c648ed43cf7c7f8a359b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544823
Reviewed-by: Felix Zhu <fcz@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>