[debugger] Decode structures returned by value.
Decodes most normal structures returned by value. ARM is simple enough
that all structures, classes, and unions returned by value are decoded.

x64 is more complicated and only a portion of the assignment algorithm
is implemented. Bitfields, vector data, and certain types not used in C
are not supported and the debugger gives up decoding these. However, all
normal structures of numbers, pointers, and other structures are
supported.

x64 floating-point register decoding is fixed. The previous code handled
a 4-byte float as a 8-byte xmm0 value and did a numeric cast. This is
not necessary and 4-byte floats are just returned in the low bits of
xmm0.

Bug: 5720

Change-Id: If7f79786a25ad9c59b1772c7833e2eb1ee8a0815
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542806
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>