[gidl] Add v2 version support in GIDL
This CL adds the v2 version in GIDL and defines GIDL bytes for all cases
that don't include a table or a union.

The idea is that only tables and unions will change in the upcoming wire
format transition, so other cases can be defined today as they won't
change.

Change-Id: I160d378929f3207dd2131a8a3446a8d67a10a3e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545642
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>