[sdk] Recreate pointer injector upon encountering error
This change adds an onError callback to PointerInjector. It is called
when pointer injector encounters error during registration or injection
of pointer events.

Also note that the pointer injector is now registered on the first
pointer down event, instead of the unreliable viewStateChanged callback
it used earlier. This should make it more robust against runtime
dependent viewStateChangedEvents.

A patch of this cl was tested in google3 by @mikaelpessa.

Test: Updated tests to reflect change in functionality.
Bug: 78992, b/181568657
Fixed: 69607
Change-Id: I26a649ca0ec8ede69aa31af74fba7abc0bc01072
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544990
Reviewed-by: David Worsham <dworsham@google.com>
Reviewed-by: Mikael Pessa <mikaelpessa@google.com>
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>