[docs] Delete stale example
Looks like after a while coverage info in Gerrit is lost. Maybe the
backing store for this info is TTL'ed? Either way, this link points to a
change that no longer shows incremental coverage on the linked patchset
or on any other patchset, so it's not a useful example.

Change-Id: Ied71ae1bef3e827a43e81adf081f780f083eeedf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546121
Commit-Queue: Shai Barack <shayba@google.com>
Fuchsia-Auto-Submit: Shai Barack <shayba@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>