[starnix] Use single_run for starnix manager children
This change updates the starnix manager to use the newly introduced
`SingleRun` collection durability for the components it runs. This means
that the children are started when they are created (don't need to bind
to them to start them), and they are automatically deleted when they are
stopped (e.g., when they exit).

Change-Id: If3662daf33db3a57c4cea8907af6cc4b6f63901b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540323
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Barth <abarth@google.com>