[build] Remove unused config-data declaration.
The config-data directory is currently defined as used but not routed to
system-metrics-logger.cm from core.cm. This means if system-metrics-logger
attempted to access /config/data in a given build it wouldn't be able to read
the directory.

Currently it appears that the code doesn't use /config/data so it
appears to be a no-op to remove the usage.

This is being fixed as part of an effort to enable static capability
flow analysis into the build.

Context: https://fuchsia-review.googlesource.com/c/fuchsia/+/545001
Bug: 61862
Change-Id: I893888305f4ff068c2d50737e0bd3229f90b978f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545954
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>