[zircon] s/ZX_INFO_PROCESS_V2/ZX_INFO_PROCESS in general cases
For general uses of the process info API, we drop the "_V2" suffix, as
they are both aliases of one another. We keep the mention of v2 only at
sites that need to be aware of both the v1 and v2 variants.

Bug: 30751
Change-Id: Ia6cc88ad8c10564e8c9374b706d6686038b32a7b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546261
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>