[brcmfmac] Refactor and rename brcmf_cfg80211_change_iface()
Refactored and renamed brcmf_cfg80211_change_iface() to
brcmf_cfg80211_change_iface_mac_role() to reduce the ambiguity
between IF type and mac role in the syslog. IF type seems to
be associated with a model in the firmware whereas mac role is a FIDL
abstraction.

The previous code used a switch statement and flags to trigger if-else
branching. The refactor uses only a switch statement instead.

Test: None. Pure refactor and logging change only.
Bug: 78905
Change-Id: Id8d69361b4d768393eadaf10b39bdfa27cc3d780
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544769
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Joshua Conner <joshconner@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>