[debugger] Support DWARF default location expressions.
DWARF 5 added a new location description format which included the
concept of a default location.

Add support for reading and using these in variable locations.

Move the existing "always valid" location to use this new concept.
Previously it had a hardcoded special-case of (0,0) for "always valid".
This required updating a number of users to avoid (0.0) and provide a
way to specify the default.

Convert the VariableLocation address range from begin/end to use the
AddressRange structure. Update users.

Bug: 77316
Change-Id: I83e34699d8d7a07d5bc89947cb51df86c600bdef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545959
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>