[ffx][config] Fix flakes when saving ffx config
Previously, multiple tests that manipulate ffx_config could racily
modify the underlying config file, which would cause the tests to flake
out. This patch fixes this issue by:

* Changing `Environment` to grab the global lock whenever the
environment file is saved or loaded.
* Changing `Environment` to atomically write to the environment file.

This change isn't the most efficient way to solve this problem, since we
only really need to lock the environment when we're reading and writing
to the same file, but we typically only encounter that in testing, so
the overhead of the global lock shouldn't impact non-testing use cases.

Change-Id: Id9d251656b30560352ccd7656b3a6387222f831b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546181
Reviewed-by: Jordon Wing <jwing@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>