[zircon][x86] map all possible segment selectors.
This CL modifies both the LDT and GDT setup code so that all possible
segment selectors are mapped.

The CL disables the LDT in x86_init_percpu by calling DisableLdt(),
which calls lldt with a null segment selector. All segment selectors
that refer to the LDT will fail with a general protection fault.

The CL also maps all the GDT entries. In zircon, the GDT limit is set to
the maximum possible value (size 0x10000), but only the first pages of
this are mapped, causing a segment selector to the end of the table to
cause a supervisor level page fault. The code now maps all the remaining
entries as copy-on-write, so future write accesses would cause a page
fault.

Before this CL, userspace could craft a segment selector that would
result in a supervisor page fault, causing the system to panic.

There are two set of tests added: a kernel unit test that tries to
access every page on the gdt, to make sure that it doesn't fault, and
userspace tests that try using selectors and make sure that the kernel
doesn't crash.

Bug: 79060
Change-Id: I6f313655d4e2f37692365c7d4c765dee04958502
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545006
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>