[kernel][vm] Multi-generation LRU Stage 1
This is the first step in the implementation towards using a
multi-generational LRU queue for the page queues. The goal of this
change is to have a lock-free path for marking pages as accessed, but
without changing too much of the page queues API.

For not changing the page queues API this means that there is still a
fixed cadence to incrementing the MRU gen, and the idea of peeking the
pager backed queue is emulated, instead of having the LRU processing
directly perform eviction.

Change-Id: Ie3d6793dd3c36539feb9481acc9a3bcb219e11d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515943
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>