[ffx] initial support for socket based diagnostics
This the change to the diagnostics bridge FIDL protocol to add support
for sending non-truncated data over a socket. This is the minimal change
to add that support but keep the implementation/functionality of logging
the same today. In a subsequent change the plan is to make use of the
new input parameters as well as actually add support for sending data
over the socket.

Change-Id: I3d940545ccce528c486ab8d23ff801e14a8a149a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528681
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>
API-Review: James Tucker <raggi@google.com>