[virtualization][vmm] Convert ReadBar/WriteBar methods to callbacks.
Currently when a PciBar is read or written to, the PciBar class
redirects the read/write to the ReadBar/WriteBar method on the parent
device. The parent device reads which BAR index was read/written, and
executes the appropriate action.

The current setup makes it difficult to allow different BARs to be
routed to different locations (e.g., one BAR for MSI-X interrupts to
the base PciDevice class, another to the derived class), and relies
on statically-allocated identifiers to determine which BAR was
read from / written to.

We update PciBar to instead take a callback for reads and writes,
allowing reads/writes to be directly routed to the appropriate handler.

Bug: 78022
Change-Id: I4ee87e6d6301c5333fd281240405a73d5f6626d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539542
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>