[pm] reorder sub-tests to attempt to avoid flake
The "writing same version doesn't send an event" sub-test occasionally
races with the asynchronous setup of the event listener and server state
such that, although the repository exists at version 1, that is not yet
observed when the write happens. That results in a spurious message
being produced into the events channel.

This change moves the test after the updated metadata is already
written. This guarantees we've tested an update already (our event queue
should therefore be empty). It's still possible to flake if an event
doesn't arrive within 1ms, but that kind of flake was already possible
in these tests (and doesn't appear to be what we actually hit).

Fixed: 78682
Change-Id: I742a9362d488ae933f08961cc59fcd865ba282d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546241
Reviewed-by: Dan Johnson <computerdruid@google.com>
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>