[fs] Make easier to identify thread safety of Watchdog class.
Was already thread safe as implemented, this just rounds up as many
members as possible to be either protected by the main lock, marked
const, or moved to be method local variables instead. The remaining
exceptions are the sync_completion_t and the std::thread. Both are
types that are meant to support thread safety and have a narrrow usage
to provide easier auditing.

Bug: 67659
Change-Id: Ia9791f1e76b6bcfada23d47f8826b7e78ecad577
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546280
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>