[next] Workaround linux app focusability
Requesting focus on linux wayland based app is broken. This results in
the shell not updating its UI state, which is built around using Scenic
focus chain as the ground truth. Linux apps do not show up in the focus
chain when request focus fails on them.

This workaround is to immediately update the local UI state, assuming
that the requestFocus succeeded.

Bug: 79463
Change-Id: I8ff540872d86344276e7700ea776f6a7343633cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/546825
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>