[run-test-suite] Remove WriteLine trait
Updates the ANSI filter so that it no longer relies on the custom
WriteLine trait. The WriteLine trait forced us to handle all artifacts
as if they are line delimited. With this change, only artifacts that
need to be filtered (stdout and stderr) need to be line delimited.

Fixed: 75134
Bug: 75133

Testing: new and existing unit tests

Change-Id: I58b82941269b65f28fa702137aa33a1a7b60a1ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545691
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>