[scenic][ui-input] Only collect contenders up to the target
Currently for TOP_HIT_AND_ANCESTORS_IN_TARGET we collect contenders
all the way up to the context. This is incorrect. This CL changes it so
that we only collect contenders in the Target subgraph.

Also updates the DispatchPolicyTest to a proper unit test and to use
fuchsia.ui.pointer, so it can be used to catch this issue.

Fixed: 79454
Bug: 72919, 53634
Change-Id: Ie4f593ba98c66d7158888e43a6976b696ca48c8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546550
Commit-Queue: Mikael Pessa <mikaelpessa@google.com>
Reviewed-by: Jaeheon Yi <jaeheon@google.com>