[diagnostics][triage] Unify multi-select and Option()
A "select:" entry with multiple selectors will now discard Missing
errors, the same as Option() does. This means that a selector
with a component moniker that's not found will no longer cause
a Missing error to be propagated, as long as a later selector
in the list finds a component.

Test: $ fx test //src/diagnostics
Fixed: 79081
Change-Id: I18f6d175fe41289e76032e9c82e6d7681e6b8838
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545705
Commit-Queue: Chris Phoenix <cphoenix@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>