[core][storage] Align product config for v1 and v2 storage restriction
Before this change, core.cml defined 3 storage capabilties, and the
persistent storage capability was always restricted to components listed
in the component ID index. With this change, product builds may choose
to restrict or unrestrict persistent storage.

This CL aligns the persistent storage restriction we have on v2
components (restrict all builds) to be the same as v1
components (only restrict on certain builds).

Concretely, here are the mechanical changes to accomplish this:
* Peel off storage capabilities defined in core.cml into shards: one
  variant restricts persistent storage, the other doesn't.
* Teach core_realm() build template a restrict_persistent_storage
  argument, which will include the appropriate shard.
  - restrict_persistent_storage is optional (default=true) for now to
  accomodate out-of-tree uses of core_realm(). Will be required in a
  follow-up change.
* For products where we don't restrict persistent storage for v1
  components, do the same for v2 components. For example, these include
  guest, overnet, recovery, terminal.

Bug: 78520
Change-Id: Iaf06286ce773a9a2387304ecad60313b6df86480
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546944
Reviewed-by: Justin Mattson <jmatt@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Vardhan Mudunuru <vardhan@google.com>